refactor: working prompt for Trivy and template with .m2 directory
Determined behavior for prompts in the generator command. Added trivy and corrected template. Each
change to the template and generator needs an npm link generator-jhipster in the project. sigh

Initial writing of .m2/settings.xml

Need to pull out Snyk and Trivy choices to sub prompts for greater fidelity in the generator by
platform.